Start prototyping smarter search terms
If you have society/website, hack/website you can jump to the first one
with j soc/web and to the second one with j hack/web.

If you happen to enter a term with different casing, we no infer it and
no longer do case insensitive searches all the time.