Add a really basic fish completion support
Its not as nice as I want it to be, cause I still can't figure out how
to pass the arguments to be in the completion.

Say, if I type `j bk<TAB>`, I want to pass `bk` to `jump hints bk`.
Still don't know if jump can even do that.