Drop the hint count limit in favour of a smart match
A smart match is a simple heuristic that the completion shell function
can use to yield better results in environments where someone sorts the
completion lists outside of the completion writer support.

That was basically a fancy way to say, that now I can hint zsh
completions better based on the number of characters the users input. I
have tweaked by heart, and there is no science or data behind it. It
seems to work okay for me, but it may not for you.

We can tweak it, future user, don't be frustrated. Leave me an issue and
we can talk.