Treat the space between multiple arguments to j as separators
This is a small, but pretty powerful change. The following
inputs are treated as the same:

```
$ j gs/assert
$ j gs assert
```

It may not seem as much, but typing a space is much easier than an os
separator. It's also easier to discover the deep search functionality.

A nice way to discover on of them: "Oh! It does _that_" moments. 🍻