Introduce full-path search support
This search is a bit hidden, but will make implementing helper functions
like autojump's `jc` only with shell.

The gist is, if there is `*` in the search term, don't search by base
name, but by the absolute path. Be sure to quote it, as it is usually
expanded by the shells.