Boost the score on bad matches
This lets `j` without arguments hint jump that the next match
may be more important than the first and boost its score a bit, so
eventually, it can overpower the one on the top of the list.

Now, this is a pretty lame approach as the factor we jump by is static
and is chosen by heart rather than mind and data... but it's a start!

We may end up in ML territory here...