Support shell-significant chars in directories during auto-completion
If you have a directory with shell-significant characters, when you
tried to auto-complete it, the text of the directory name will be
inserted in the shell and you may start a subshell in the worst case, or
end up in a bad directory in the better case. This can be a security
concern as well.

For example, if you have the following directory:

```
/Users/genadi/Test/test folder with ` and - " all that jazz
```

`j jazz<TAB>`

Will auto-complete the line to...

```
j /Users/genadi/Test/test folder with ` and - " all that jazz
```

If you press Enter at that stage in Bash, you'll end up with:

```
bash-3.2$ j /Users/genadi/Test/test folder with ` and - " all that jazz
>```

Waiting for you to close the ` and execute the command. Not good!

This commit escapes the directory name when you try to auto-complete it. Mind
that not skipping the auto-completions works fine, so `j jazz` would have
still landed you in the proper directory.