Fix `j rails web` for fish shell
In the fish shell integration we used to quote the arguments passed to
`j` and `j rails web` was passed as `jump cd "rails web"`. "rails web"
is treated as a single argument to `jump cd`. We want the multiple
arguments, however, as we have builtin mechanism to either join the
arguments with OS separators for deep search, or to ignore them for
`z`-like compatible, a former friend a needed. I say former, because he
uses zoxide now. Cancellable offence!

I don't hold grudges, tho.