Initial basedir jumping support
I got the request to support jumping inside a base directory. This can
be a monorepo git project, a perforce workspace and the likes.

To initiate a base directory jump you need to specify a absolute path to
the base directory as the first argument to `j`, the second to last
arguments can be any jump search term. The base directory needs to be
accessible for the base mode to initiate

```
$ j /home/genadi/work/dext app
$ pwd # /home/genadi/work/dext/application
```

In base mode, jumps will be restricted to directories contained within
the base path. If a search term is an accessible directory inside the
base, jump will enter it like cd and avoid a search. This is a contained
jump.

Now, using `j` is all about saving keystrokes and this is anything but
that. That's why you would want to use this as a base to another shell
function, say `z`:

```
z() {
  j "$(git rev-parse --show-toplevel)" $@
}
```

`z` will look for a git repo root dir and set it as a base dir for
jumps. Pretty neat. I call this the j(ay)-z mode.

I am running into a philosophical usage conundrum, though. I always
wanted `j` to be the only entry point for jump and to not expose any flags to it.
All the modes are inferred from the search term. I like it, but I can see
how it is confusing to others.

I need to do a better job at documenting this or cave in and introduce
options. I can leave the options to the `jump` command, though.

I want to improve the implementation itself too, as it's a bit rough at
the moment, but I also want to test this. Will iterate on the jumping
internals before our next release.