Inline the based mode logic into a single cd function
As a side effect, we have lifted the CD mode outside of based mode
completion and we trigger it with a trailing slash in the input.

For example:

```bash
$ j app/views/
```

Will go to the children `app/views/` directory if it exists. No need for
the `j . app/views/` hint. This effectively makes `j` act like `cd`
should the user want to use it like this. I'll work on adding the
current directories in the autocompletion hints down the line.